+++ include/berry.h
@@ -25,7 +25,6 @@ void ObjectEventInteractionPickBerryTree(void);
 void ObjectEventInteractionRemoveBerryTree(void);
 void ObjectEventInteractionApplyMulch(void);
 bool8 PlayerHasBerries(void);
-void SetBerryTreesSeen(void);
 bool32 BerryTreeGrow(struct BerryTree *tree);
 
 extern const struct Berry gBerries[];