+++ CHANGELOG.md
@@ -1,19 +1,33 @@
 # Re:Emerald Changelog
 
 ## In-development
+**Update 1**
 - Modified the in-game trades to be more useful
+
+**Updates 2-6**
 - Implemented the Swarm mechanic:
     - Swarm Pokemon show up in the Pokedex correctly
     - Implemented Birch's book for active swarms
 - Increased the variety of wild Pokemon in some areas
+
+**Update 7**
 - Added a list to the Pokemon's menu that shows any field moves it can use that can be used from the overworld.
     - This fixes the issue that made Regirock unobtainable due to being locked behind Rock Smash.
+
+**Update 8**
 - Changed the line of sight of the Gentleman in Route 121. Now you can reach the item without fighting him or without going through the grass
 - Shortened the messages when picking up a berry.
 - When asked to change Pokemon, you can still see what Pokemon the opponent will send out.
 - Fixed the capitalization and name of some items and descriptions.
+
+**Update 9**
 - Expanded the bag to hold 1 of every item at the same time.
 - Changed Exp. Share to work like in Gen VI
+
+**Update 10**
+- Berry plants no longer disappear or start growing again after raching the fully grown stage
+- The formula for calculating the extra berry yields was changed and is no longer random
+- The minimum and maximum yield for all berries were changed
 
 ## 0.1.0
 - Initial game release